+++ native_extension/VisionFaceANE/ane/build_multi.sh
@@ -4,7 +4,7 @@
 pathtome=$0
 pathtome="${pathtome%/*}"
 
-PROJECTNAME=VisionFaceANE
+PROJECTNAME=FaceANE
 libSuffix="_LIB"
 
 AIR_SDK="/Users/eoinlandy/SDKs/AIRSDK_33"
@@ -59,7 +59,7 @@ cp "$pathtome/../../../native_library/android/FirebaseANE/VisionFace/build/outpu
 echo "getting Android jars"
 unzip "$pathtome/platforms/android/app-release.aar" "classes.jar" -d "$pathtome/platforms/android"
 unzip "$pathtome/platforms/android/app-release.aar" "res/*" -d "$pathtome/platforms/android"
-mv "$pathtome/platforms/android/res" "$pathtome/platforms/android/com.tuarua.firebase.$PROJECTNAME-res"
+mv "$pathtome/platforms/android/res" "$pathtome/platforms/android/com.tuarua.mlkit.vision.$PROJECTNAME-res"
 
 #Run the build command.
 echo "Building ANE."
@@ -71,15 +71,15 @@ echo "Building ANE."
 -platform default -C "$pathtome/platforms/default" "library.swf" \
 -platform Android-x86 \
 -C "$pathtome/platforms/android" "library.swf" "classes.jar" \
-com.tuarua.firebase.${PROJECTNAME}-res/. \
+com.tuarua.mlkit.vision.${PROJECTNAME}-res/. \
 -platformoptions "$pathtome/platforms/android/platform.xml" \
 -platform Android-ARM \
 -C "$pathtome/platforms/android" "library.swf" "classes.jar" \
-com.tuarua.firebase.${PROJECTNAME}-res/. \
+com.tuarua.mlkit.vision.${PROJECTNAME}-res/. \
 -platformoptions "$pathtome/platforms/android/platform.xml" \
 -platform Android-ARM64 \
 -C "$pathtome/platforms/android" "library.swf" "classes.jar" \
-com.tuarua.firebase.${PROJECTNAME}-res/. \
+com.tuarua.mlkit.vision.${PROJECTNAME}-res/. \
 -platformoptions "$pathtome/platforms/android/platform.xml" \
 
 #remove the frameworks from sim and device, as not needed any more
@@ -90,5 +90,5 @@ rm -r "$pathtome/platforms/ios/device"
 rm -r "$pathtome/platforms/default"
 rm "$pathtome/$PROJECTNAME.swc"
 rm "$pathtome/library.swf"
-rm -r "$pathtome/platforms/android/com.tuarua.firebase.$PROJECTNAME-res"
+rm -r "$pathtome/platforms/android/com.tuarua.mlkit.vision.$PROJECTNAME-res"
 echo "Finished."